* Rename OakRReference to OakRKeyBufferImpl (removed the old, unused OakRKeyBufferImpl) * Rename OakUnsafeRef to OakUnsafeDirectBuffer * Add proper internal documentation in OakUnsafeDirectBuffer
