+++ core/src/main/java/com/oath/oak/OakWBufferImpl.java
@@ -13,7 +13,7 @@ import java.nio.ByteOrder;
 
 // An instance of OakWBufferImpl is only used when the write lock of the value referenced by it is already acquired.
 // This is the reason no lock is acquired in each access.
-public class OakWBufferImpl implements OakWBuffer, OakUnsafeRef {
+public class OakWBufferImpl implements OakWBuffer, OakUnsafeDirectBuffer {
 
     private ByteBuffer bb;
     private int dataPos;
